# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NachtLabs",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/auth/setup": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Setup",
        "operationId": "setup_api_v1_auth_setup_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Setup"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOutput"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Login",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Login"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginOutput"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/api/v1/auth/mfa/verify": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Mfa Verify",
        "operationId": "mfa_verify_api_v1_auth_mfa_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MFAChallenge"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOutput"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": []
      }
    },
...
    "/api/v1/auth/reauthenticate": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Reauthenticate",
        "operationId": "reauthenticate_api_v1_auth_reauthenticate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Reauthenticate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "type": "object",
                  "title": "Response Reauthenticate Api V1 Auth Reauthenticate Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/mfa/confirm": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Mfa Confirm",
        "operationId": "mfa_confirm_api_v1_auth_mfa_confirm_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MFACode"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "items": {
                      "type": "string"
                    },
                    "type": "array"
                  },
                  "type": "object",
                  "title": "Response Mfa Confirm Api V1 Auth Mfa Confirm Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/forgot-password": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Forgot",
        "operationId": "forgot_api_v1_auth_forgot_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Forgot Api V1 Auth Forgot Password Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/api/v1/auth/reset-password": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Reset",
        "operationId": "reset_api_v1_auth_reset_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordToken"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "type": "object",
                  "title": "Response Reset Api V1 Auth Reset Password Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/api/v1/auth/accept-invitation": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Accept",
        "operationId": "accept_api_v1_auth_accept_invitation_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordToken"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "type": "object",
                  "title": "Response Accept Api V1 Auth Accept Invitation Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/api/v1/auth/preferences": {
      "put": {
        "tags": [
          "Authentication"
        ],
        "summary": "Preferences",
        "operationId": "preferences_api_v1_auth_preferences_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Preference"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Preferences Api V1 Auth Preferences Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EmailInput": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "email"
        ],
        "title": "EmailInput"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Login": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "format": "password",
            "title": "Password",
            "writeOnly": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "Login"
      },
      "LoginOutput": {
        "properties": {
          "mfa_required": {
            "type": "boolean",
            "title": "Mfa Required"
          },
          "challenge": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Challenge"
          },
          "user": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserOutput"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "mfa_required"
        ],
        "title": "LoginOutput"
      },
      "MFAChallenge": {
        "properties": {
          "challenge": {
            "type": "string",
            "format": "password",
            "title": "Challenge",
            "writeOnly": true
          },
          "code": {
            "type": "string",
            "maxLength": 64,
            "minLength": 6,
            "format": "password",
            "title": "Code",
            "writeOnly": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "challenge",
          "code"
        ],
        "title": "MFAChallenge"
      },
      "MFACode": {
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 64,
            "minLength": 6,
            "format": "password",
            "title": "Code",
            "writeOnly": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code"
        ],
        "title": "MFACode"
      },
...
      "PasswordToken": {
        "properties": {
          "token": {
            "type": "string",
            "format": "password",
            "title": "Token",
            "writeOnly": true
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 12,
            "format": "password",
            "title": "Password",
            "writeOnly": true
          },
          "name": {
            "type": "string",
            "maxLength": 120,
            "title": "Name",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "token",
          "password"
        ],
        "title": "PasswordToken"
      },
...
      "Preference": {
        "properties": {
          "theme": {
            "type": "string",
            "enum": [
              "midnight",
              "graphite",
              "canvas",
              "forest",
              "nordic",
              "solarized",
              "system"
            ],
            "title": "Theme"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "theme"
        ],
        "title": "Preference"
      },
...
      "Reauthenticate": {
        "properties": {
          "password": {
            "type": "string",
            "format": "password",
            "title": "Password",
            "writeOnly": true
          },
          "code": {
            "anyOf": [
              {
                "type": "string",
                "format": "password",
                "writeOnly": true
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "password"
        ],
        "title": "Reauthenticate"
      },
...
      "Setup": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 12,
            "format": "password",
            "title": "Password",
            "writeOnly": true
          },
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "default": "Owner"
          },
          "organization": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Organization",
            "default": "NachtLabs"
          },
          "bootstrap_token": {
            "anyOf": [
              {
                "type": "string",
                "format": "password",
                "writeOnly": true
              },
              {
                "type": "null"
              }
            ],
            "title": "Bootstrap Token"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "Setup"
      },
...
      "UserOutput": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "role": {
            "type": "string",
            "enum": [
              "owner",
              "admin",
              "operator",
              "contributor",
              "viewer"
            ],
            "title": "Role"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          },
          "theme": {
            "type": "string",
            "enum": [
              "midnight",
              "graphite",
              "canvas",
              "forest",
              "nordic",
              "solarized",
              "system"
            ],
            "title": "Theme"
          },
          "mfa_enabled": {
            "type": "boolean",
            "title": "Mfa Enabled"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "name",
          "role",
          "active",
          "theme",
          "mfa_enabled",
          "version"
        ],
        "title": "UserOutput"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "nachtlabs_session",
        "description": "Mutations also require Origin and X-CSRF-Token."
      },
      "ScopedAPIKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "Project-restricted service-account key."
      }
    }
  },
  "security": [
    {
      "SessionCookie": []
    },
    {
      "ScopedAPIKey": []
    }
  ]
}